+++ src/objects/Sphere.cpp
@@ -11,7 +11,7 @@
 
 #include "objects/Sphere.hpp"
 
-Sphere::Sphere(const vec3& _center, double _radius,
+Sphere::Sphere(const vec3& _center, float _radius,
                std::shared_ptr<material> _mat)
     : center(_center), radius(_radius), mat(_mat) {}
 
@@ -19,7 +19,7 @@ Sphere::~Sphere() {}
 
 vec3 Sphere::get_center() { return center; }
 
-double Sphere::get_radius() { return radius; }
+float Sphere::get_radius() { return radius; }
 
 bool Sphere::intersect(const ray& r, const float t_min, const float t_max,
                        hit_record& rec) {
